backend: Reduced number of queries
Instead of making one (or even two!) queries per element in a list to
fill fields that are in a different collection, build a list of ids and
use the $in operator to retrieve all of them in one command.
After that, a matching phase (double nester for, for sake of simplicity)
fills the requested fields.